backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic==0.39.0
google-search-results==2.4.2
pydantic==2.9.2
python-multipart==0.0.12
